Extra symlink fix. Check for char count.
Fixed the error with how many symlinks are generated. Checks first to
see if symlinks already exist. Adds in a check for how many characters
are used in the qshields command (forecasting to the worst-case
scenario) and throws a warning or error depending on how close to the
count it.